Integrating GPT and Mistral code in config.py
